tests/wpt/web-platform-tests/html/browsers/windows/browsing-context-names/choose-_parent-002.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: Choose browsing context - '_parent' (nested contexts)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(t => {
  var topWindow;
  t.add_cleanup(() => topWindow.close());
  window.addEventListener('message', t.step_func_done(e => {
    assert_equals(e.data.name, 'iframeParent');
    assert_false(e.data.isTop, 'window.parent is not top');
  }));
  topWindow = window.open('resources/choose-_parent-002-window.html', '_blank');
}, 'choosing _parent context: multiple nested contexts');
</script>
